Use originating font for highlight pseudo units
Per CSS WG resolution, CSS Highlight Pseudo elements should
use the font from their originating element when computing
units for text decorations and the like.

To support this, a new length storage class is defined,
LazyEvaluatedFontLength, to store information to support
length unit evaluation at use time based on the originating
element. Note we must resolve at use time because the same
highlight may be used with multiple elements.

Text decoration properties are changed to make use of this
new unit. We ignore shadows for now because we do not enable
them in highlight pseudos. The text decoration painting
code (and ink overflow) is updated to make use of the new
information.

Bug: 1466288
Change-Id: Ice191b44c3420f1b4c4a9019d879c92239d5f299