sensors: Add content_shell code path for new sensor manipulation commands.
content_shell does not use ChromeDriver, so we need a separate
implementation for the testdriver.js commands we are adding.

This implementation is anchored on Blink's Internals system, which then uses
Mojo to call the virtual sensor-related Mojo commands in the SensorProvider
interface. Once the calls reach content/browser, the code path is identical
to what happens via ChromeDriver and CDP.

Bug: 1278377
Change-Id: I13f49d529b36916d1f98b4792839e818fd7e8bdb