sensors: Add support for virtual sensors in //services
This implements the concept of "virtual sensors" from the Automation
section of the Generic Sensor spec. From a spec perspective, virtual
sensors act like device sensors whose readings and properties are
controlled by users.

In Chromium, virtual sensors are represented by the
VirtualPlatformSensor class. There is a specific provider for them,
VirtualPlatformSensorProvider.

There is one VirtualPlatformSensorProvider instance per Mojo connection
that calls CreateVirtualSensor(), so that virtual sensors created in one
page (not frame) do not interfere with other pages (virtual sensors of
the same type may be used with completely different properties,
constraints, and readings in different pages and they should not affect
each other). Additionally, while regular PlatformSensorProvider
instances essentially behave like singletons (SensorProviderImpl's
lifetime is tied to DeviceSensor's), VirtualPlatformSensorProvider
instances are tied to that of the Mojo endpoint that calls
CreateVirtualSensor() -- in other words, a SensorProviderProxyImpl in
//content (which exists on a per-Page basis).

One downside of this design is that the virtual sensor-related
operations need to be added to the SensorProvider Mojo interface, and
consequently the new methods need to be at least stubbed by all
implementations of the Mojo interface in the tree.

SensorProviderImpl is responsible for managing virtual and real sensors.
Its `provider_` member remains the same and manages real sensors. The
new `virtual_providers_` member maps a mojo::ReceiverId to a
VirtualPlatformSensorProvider. GetSensor() calls first try to find a
virtual sensor provider and fall back to the real one otherwise.
Therefore, real sensors and virtual sensors are unaware of each other.

One consequence is that fusion sensors will continue to attempt to use
only real sensors as sources, unless the fusion sensor itself is created
as a virtual sensor.

Another property of this design is that it is possible for real and
virtual sensors of the same type to coexist in the same page: real
sensors are used as long as CreateFakeSensor() is not called; once that
happens new sensors will be virtual, but existing ones will not be
affected.

Finally, the new code is tested by making GenericSensorServiceTest use
virtual sensors. GenericSensorServiceTest already tested the behavior of
the SensorImpl and SensorProviderImpl layer, so it is the ideal place to
start testing how SensorProviderImpl handles virtual sensors. It also
allows us to test a lot of VirtualPlatformSensorProvider directly.

This also helps with bug 1427302 now that usage of
FakePlatformSensor{Provider} usage has reduced considerably.

Bug: 1278377, 1427302
Change-Id: Iebe3e70bfe1d154e305243f876d9ae8b436757c9