Move orientation-event web tests from wpt_internal to external/wpt
The orientation-event web tests require a method to minimize and
restore the browser window. Initially, this functionality was absent in
the WPT. With the addition of these services to the testdriver.js, the
orientation-event web tests can now be moved to the WPT.

As orientation-event was removed the redundant sensor-helpers.js file
could be removed, which was duplicated from generic_sensor_mocks.js.

From WPT page-visibility-manual.https.html was removed from orientation
and motion sensor tests. This test needed manually to switch the page
to the background, then switch back to foreground. Thanks to
testdriver.js manual operations are not needed anymore.

Bug: 1171539
Change-Id: I5bcb1d4236d05829e88e421a8c464b9a75dd0739