Locate parent fragment of nested OOF multicol correctly.
The code didn't sufficiently examine the containing block chain. We need
to locate the nearest non-out-of-flow containing block that has a
non-out-of-flow fragmentation context root as *its* containing block
(unless the fragmentation context root is the LayoutView (printing),
which is hard-coded as absolutely positioned) - and then place the new
fragment under this fragmentation context root. This is what the
fragment tree looks like, so, when replacing a fragment after OOF
layout, it needs to behave exactly the same, or we'll find to find the
old fragment in the fragment tree.

Bug: 1459154
Change-Id: I0a3ae3ea07331db13566883f98dee3f1a989e5c4